base: Use a system-wide recents corner radius
With Google implementation of quickstep, the rounded corners are set indivdually according to the view.
If it's the AOSP default recents, it'll take a value for rounded corners.
If it's Launcher3 or any other app, it'll take a custom value that's declared on the app.

Let's make a public value and use it on both.

- Include
https://github.com/PixysOS/frameworks_base/commit/aa905b2eec4b3038cb16a3c70ea734db65d3f398#diff-c3df04adca74dc8a9d525119e0a107c1

Change-Id: Ic91e3dd631f0466e1c6ff68755b1808fe698533d